Fix processing Mbox messages with a foreign encoding
A Mbox file may contain messages in various encoding, not necessarily
the one the user is currently using.  For this reason, it is advisable to
read the Mbox file as a binary file (that is with a ASCII8BIT encoding),
and attempt to adjust the encoding for each message in turn.

We first try to force the encoding to UTF-8 (which also accept ASCII7BIT
encoding), then ISO-8859-1 if the resulting string has not a valid
encoding.  This avoids choking on unknown encoding if non-reporting
messages are present in a report Mbox.