Add support for deduplicating reports
From time to time, a report is delivered multiple times.  When this
happen, the duplicates are generally sent a few minutes after the first
message, with the same report (same content, same report ID), and the
same Message-ID, but can come from different servers and have a distinct
DKIM signature.

Add some detection mechanism to detect if a report has already been
ingested to avoid storing multiple instances of the same events and
provide more accurate information.

When batch-importing reports, it may be necessary to adjust the settings
for duplicates detection by using a date range that cover all relevant
reports and increasing the number of reports that can be found.